LPS-106089 Move Data record collection logic to SPI module
