+++ modules/apps/data-engine/data-engine-spi/src/main/java/com/liferay/data/engine/spi/model/InternalDataRecordCollection.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.data.engine.rest.internal.model;
+package com.liferay.data.engine.spi.model;
 
 import com.liferay.expando.kernel.model.ExpandoBridge;
 import com.liferay.portal.kernel.model.ClassedModel;
